+++ src/lib/configure.ts
@@ -2,8 +2,12 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as os from 'os';
 
-//TODO: Move to XDG_CONFIG
-const configFile = path.resolve(os.homedir(), '.clubhouse-cli', 'config.json');
+const configDir =
+    process.env.XDG_CONFIG_HOME ||
+    path.resolve(process.env.XDG_DATA_HOME || os.homedir(), '.config', 'clubhouse-cli');
+
+const configFile = path.resolve(configDir, 'config.json');
+const legacyConfigDir = path.resolve(os.homedir(), '.clubhouse-cli');
 
 export interface Config {
     mentionName: string;
@@ -20,6 +24,10 @@ export interface Config {
 
 export const loadConfig: () => Config = () => {
     const envToken = process.env.CLUBHOUSE_API_TOKEN;
+    if (fs.existsSync(legacyConfigDir)) {
+        createConfigDir();
+        fs.renameSync(legacyConfigDir, configDir);
+    }
     if (fs.existsSync(configFile)) {
         try {
             const config = JSON.parse(fs.readFileSync(configFile, 'utf8'));
@@ -38,12 +46,19 @@ export const loadConfig: () => Config = () => {
     return {};
 };
 
+const createConfigDir = () => {
+    const dir = path.dirname(configDir);
+    if (!fs.existsSync(dir)) {
+        fs.mkdirSync(dir);
+    }
+    if (!fs.existsSync(configDir)) {
+        fs.mkdirSync(configDir);
+    }
+};
+
 const saveConfig = (opt: any) => {
-    const dir = path.dirname(configFile);
     try {
-        if (!fs.existsSync(dir)) {
-            fs.mkdirSync(dir);
-        }
+        createConfigDir();
         fs.writeFileSync(configFile, JSON.stringify(opt), { flag: 'w' });
         return true;
     } catch (e) {